refactor: refactor file service for app data
