+++ src/services/item/plugins/app/appData/appData.schemas.ts
@@ -9,6 +9,7 @@ import { customType, registerSchemaAsRef } from '../../../../../plugins/typebox'
 import { errorSchemaRef } from '../../../../../schemas/global';
 import { accountSchemaRef, nullableAccountSchemaRef } from '../../../../account/account.schemas';
 import { itemSchemaRef } from '../../../item.schemas';
+import { APP_DATA_TYPE_FILE } from '../constants';
 
 const appDataSchema = customType.StrictObject(
   {
@@ -131,3 +132,38 @@ export const getForOne = {
     '4xx': errorSchemaRef,
   },
 } as const satisfies FastifySchema;
+
+export const upload = {
+  operationId: 'createAppDataFile',
+  tags: ['app', 'app-data', 'file'],
+  summary: 'Create app data file',
+  description: `Upload a file to create a corresponding app data. The created app data will be "${APP_DATA_TYPE_FILE}" and visibility ${AppDataVisibility.Member}. The data property will contain the file properties.`,
+
+  response: {
+    [StatusCodes.OK]: appDataWithLegacyPropsSchemaRef,
+    '4xx': errorSchemaRef,
+  },
+};
+
+export const download = {
+  operationId: 'downloadAppDataFile',
+  tags: ['app', 'app-data', 'file'],
+  summary: 'Download app data file',
+  description: 'Download app data file.',
+
+  params: customType.StrictObject({
+    id: customType.UUID({
+      description: 'Id of the app data corresponding to the file to download',
+    }),
+  }),
+  querystring: customType.StrictObject({
+    replyUrl: Type.Boolean({
+      default: false,
+    }),
+  }),
+
+  response: {
+    [StatusCodes.OK]: Type.String({ format: 'uri' }),
+    '4xx': errorSchemaRef,
+  },
+};
